move ctid to function
